% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_image.R
\name{add_image}
\alias{add_image}
\title{Adds png image to a plot}
\usage{
add_image(obj, x = NULL, y = NULL, width = NULL, interpolate = TRUE)
}
\arguments{
\item{obj}{an image file imported as an array (e.g. png::readPNG, jpeg::readJPEG)}

\item{x}{mid x coordinate for image}

\item{y}{mid y coordinate for image}

\item{width}{width of image (in x coordinate units)}

\item{interpolate}{Logical. Passed to graphics::rasterImage. Indicates whether to apply linear interpolation to the image when drawing.}
}
\description{
Function adapted from from Marc in the box to add a png picture to a plot. Original function in: https://stackoverflow.com/questions/27800307/adding-a-picture-to-plot-in-r
}
