% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antifa.R
\name{antifa}
\alias{antifa}
\title{Creates antifa logo}
\usage{
antifa(
  x = "r-ladies+",
  bg_col = "#a70000",
  text_size = 3,
  save = FALSE,
  image_dir = NULL,
  image_name = NULL
)
}
\arguments{
\item{x}{Character. Name of the group identity}

\item{bg_col}{Background color for the flag}

\item{text_size}{Numeric. Text size for the image}

\item{save}{Logical}

\item{image_dir}{Character. Path to sabe image}

\item{image_name}{Character. Image file name without extension}
}
\description{
This function takes the antifa logo, edits the color of the flag and includes
text. It can save the png on disk and return the path to be used in the next
step, that is creating the hexSticker
}
