% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Reads antifa flag image and edits pallete of antifascist flag}
\usage{
get_image(logo_width = 480, bg_col = NULL)
}
\arguments{
\item{logo_width}{Numeric. Width (in pixels) of the image}

\item{bg_col}{Color of background}
}
\description{
This function was originally written by Danielle Navarro for package rainbowr
it takes a template .svg and edits its palette. In this case the template is
the antifascist logo with a white background. The color of the flag is red by
 default but can be editted (bg.col). The function writes png and svg to disk
 and returns an svg
}
